AddCardActivity reworked as AddEditCardActivity.
AddEditCardActivity is now the activity responsible for both adding and editing cards.
Added buttons to CardDetailActivity to edit or delete the selected card from wishlist. If card has lockstatus = true, these buttons are disabled.
CardListActivity now takes on the correct title of the activity ("My Collection" or "My Wishlist")
Login screen now uses email form entry type. Meaning it won't add spaces to the end of the text entry and use the correct keyboard layout.

The actual addition/editing of the card with AddEditCardActivity needs to be finished.